Simplifies search through relations usage
