+++ front/src/pages/companies/companies-columns.tsx
@@ -1,12 +1,4 @@
-import { CellContext, createColumnHelper } from '@tanstack/react-table';
-import {
-  Company,
-  mapToCompany,
-} from '../../interfaces/entities/company.interface';
-import { updateCompany } from '../../services/api/companies';
-import ColumnHead from '../../components/table/ColumnHead';
-import CompanyChip from '../../components/chips/CompanyChip';
-import EditableText from '../../components/editable-cell/EditableText';
+import { useMemo } from 'react';
 import {
   FaRegBuilding,
   FaCalendar,
@@ -14,126 +6,27 @@ import {
   FaMapPin,
   FaRegUser,
   FaUsers,
-  FaBuilding,
 } from 'react-icons/fa';
+import { CellContext, createColumnHelper } from '@tanstack/react-table';
+
+import { SEARCH_USER_QUERY } from '../../services/api/search/search';
+import { SearchConfigType } from '../../interfaces/search/interface';
+
+import { Company } from '../../interfaces/entities/company.interface';
+import { updateCompany } from '../../services/api/companies';
+import { User, mapToUser } from '../../interfaces/entities/user.interface';
+
+import ColumnHead from '../../components/table/ColumnHead';
+import Checkbox from '../../components/form/Checkbox';
+import { SelectAllCheckbox } from '../../components/table/SelectAllCheckbox';
+import EditableDate from '../../components/editable-cell/EditableDate';
+import EditableRelation from '../../components/editable-cell/EditableRelation';
+import EditableChip from '../../components/editable-cell/EditableChip';
+import EditableText from '../../components/editable-cell/EditableText';
 import PersonChip, {
   PersonChipPropsType,
 } from '../../components/chips/PersonChip';
-import EditableChip from '../../components/editable-cell/EditableChip';
-import { Companies_Order_By } from '../../generated/graphql';
-import {
-  SEARCH_COMPANY_QUERY,
-  SEARCH_USER_QUERY,
-} from '../../services/api/search/search';
-import EditableDate from '../../components/editable-cell/EditableDate';
-import EditableRelation from '../../components/editable-cell/EditableRelation';
-import { User, mapToUser } from '../../interfaces/entities/user.interface';
-import { useMemo } from 'react';
-import { SelectAllCheckbox } from '../../components/table/SelectAllCheckbox';
-import Checkbox from '../../components/form/Checkbox';
-import { SortType } from '../../interfaces/sorts/interface';
-import { FilterConfigType } from '../../interfaces/filters/interface';
-import { SearchConfigType } from '../../interfaces/search/interface';
-
-export const availableSorts = [
-  {
-    key: 'name',
-    label: 'Name',
-    icon: <FaBuilding />,
-    _type: 'default_sort',
-  },
-  {
-    key: 'employees',
-    label: 'Employees',
-    icon: <FaUsers />,
-    _type: 'default_sort',
-  },
-  {
-    key: 'domain_name',
-    label: 'Url',
-    icon: <FaLink />,
-    _type: 'default_sort',
-  },
-  {
-    key: 'address',
-    label: 'Address',
-    icon: <FaMapPin />,
-    _type: 'default_sort',
-  },
-  {
-    key: 'created_at',
-    label: 'Creation',
-    icon: <FaCalendar />,
-    _type: 'default_sort',
-  },
-] satisfies Array<SortType<Companies_Order_By>>;
-
-export const availableFilters = [
-  {
-    key: 'company_name',
-    label: 'Company',
-    icon: <FaBuilding />,
-    searchConfig: {
-      query: SEARCH_COMPANY_QUERY,
-      template: (searchInput) => ({
-        name: { _ilike: `%${searchInput}%` },
-      }),
-      resultMapper: (company) => ({
-        render: (company) => company.name,
-        value: mapToCompany(company),
-      }),
-    },
-    selectedValueRender: (company) => company.name || '',
-    operands: [
-      {
-        label: 'Equal',
-        id: 'equal',
-        whereTemplate: (company) => ({
-          name: { _eq: company.name },
-        }),
-      },
-      {
-        label: 'Not equal',
-        id: 'not-equal',
-        whereTemplate: (company) => ({
-          _not: { name: { _eq: company.name } },
-        }),
-      },
-    ],
-  } satisfies FilterConfigType<Company, Company>,
-  {
-    key: 'company_domain_name',
-    label: 'Url',
-    icon: <FaLink />,
-    searchConfig: {
-      query: SEARCH_COMPANY_QUERY,
-      template: (searchInput) => ({
-        name: { _ilike: `%${searchInput}%` },
-      }),
-      resultMapper: (company) => ({
-        render: (company) => company.domainName,
-        value: mapToCompany(company),
-      }),
-    },
-    selectedValueRender: (company) => company.domainName || '',
-    operands: [
-      {
-        label: 'Equal',
-        id: 'equal',
-        whereTemplate: (company) => ({
-          domain_name: { _eq: company.domainName },
-        }),
-      },
-      {
-        label: 'Not equal',
-        id: 'not-equal',
-        whereTemplate: (company) => ({
-          _not: { domain_name: { _eq: company.domainName } },
-        }),
-      },
-    ],
-  } satisfies FilterConfigType<Company, Company>,
-];
+import CompanyChip from '../../components/chips/CompanyChip';
 
 const columnHelper = createColumnHelper<Company>();
